requirements.txt
fastapi
uvicorn
python-multipart
Pillow
numpy
transformers
torch
onnxruntime
optimum[onnxruntime]
pydantic
pytest
httpx
streamlit
torchvision
datasets
accelerate
